+++ multiqc/modules/isoseq3_refine/isoseq3_refine.py
@@ -24,6 +24,7 @@ class MultiqcModule(BaseMultiqcModule):
             anchor="isoseq3_refine",
             href="https://github.com/PacificBiosciences/IsoSeq",
             info="is for poly(A) tails detection/trimming and concatemer removal.",
+            doi="",
         )
 
         self.isoseq3_refine_data = dict()
@@ -63,8 +64,8 @@ class MultiqcModule(BaseMultiqcModule):
             data_means = data[["fivelen", "threelen", "polyAlen", "insertlen"]].mean().to_frame().transpose()
             data_std = data[["fivelen", "threelen", "polyAlen", "insertlen"]].std().to_frame().transpose()
             data_max = data[["fivelen", "threelen", "polyAlen", "insertlen"]].max().to_frame().transpose()
-            data_csv_strand = data['primer'].value_counts().to_dict()
-            data_csv_primer = data['strand'].value_counts().to_dict()
+            data_csv_strand = data["primer"].value_counts().to_dict()
+            data_csv_primer = data["strand"].value_counts().to_dict()
 
             data_csv["min_fivelen"] = data_min["fivelen"][0]
             data_csv["mean_fivelen"] = data_means["fivelen"][0]
@@ -102,22 +103,22 @@ class MultiqcModule(BaseMultiqcModule):
             "title": "Number of full length reads",
             "description": "Number of CCS where both primers have been detected",
             "suffix": " fl",
-            'format': '{:,}',
-            'scale': 'spectral',
+            "format": "{:,}",
+            "scale": "spectral",
         }
         headers["num_reads_flnc"] = {
             "title": "Number of non chimeric full length reads",
             "description": "Number of non chimeric CCS where both primers have been detected",
             "suffix": " flnc",
-            'format': '{:,}',
-            'scale': 'RdYlGn',
+            "format": "{:,}",
+            "scale": "RdYlGn",
         }
         headers["num_reads_flnc_polya"] = {
             "title": "Number of poly(A) free non chimeric full length reads",
             "description": "Number of non chimeric CCS where both primers have been detected and the poly(A) tail has been removed",
             "suffix": " flnc_polya",
-            'format': '{:,}',
-            'scale': 'spectral',
+            "format": "{:,}",
+            "scale": "spectral",
         }
         self.general_stats_addcols(gstats_data, headers)
 
@@ -126,53 +127,36 @@ class MultiqcModule(BaseMultiqcModule):
         self.add_section_insertlen()
         self.add_section_polyalen()
         self.add_section_threelen()
-        # self.add_section(
-        #     name="Full length CCS report",
-        #     anchor="fl-report",
-        #     description="Detail the CCS selected and trimmed by isoseq3 refine",
-        #     helptext="""
-        #              _fl / Full Length_
-        #              * Both primers have been detected and trimmed from CCS
-        #              _flnc / Full Length Non Chimeric_
-        #              * Both primers have been detected and trimmed from CCS
-        #              * No chimeric structure has been detected (ex: primer inside the CCS)
-        #              _flnc_polya / Full Length Non Chimeric no poly(A) tail_
-        #              * Both primers have been detected and trimmed from CCS
-        #              * No chimeric structure has been detected (ex: primer inside the CCS)
-        #              * The poly(A) tails has been detected and trimmed
-        #             """,
-        #     plot=table.plot(self.isoseq3_refine_data_json),
-        # )
 
     def add_section_fivelen(self):
         plot_data = dict()
         for s_name, data in self.isoseq3_refine_data_csv.items():
             plot_data[s_name] = dict()
-            plot_data[s_name]['min_fivelen'] = data['min_fivelen']
-            plot_data[s_name]['mean_fivelen'] = data['mean_fivelen']
-            plot_data[s_name]['std_fivelen'] = data['std_fivelen']
-            plot_data[s_name]['max_fivelen'] = data['max_fivelen']
+            plot_data[s_name]["min_fivelen"] = data["min_fivelen"]
+            plot_data[s_name]["mean_fivelen"] = data["mean_fivelen"]
+            plot_data[s_name]["std_fivelen"] = data["std_fivelen"]
+            plot_data[s_name]["max_fivelen"] = data["max_fivelen"]
 
         headers = OrderedDict()
-        headers['min_fivelen'] = {
-            'title': "Min 5' primer length",
-            'description': "The minimum 5' primer length in base pair",
-            'scale': 'spectral',
+        headers["min_fivelen"] = {
+            "title": "Min 5' primer length",
+            "description": "The minimum 5' primer length in base pair",
+            "scale": "spectral",
         }
-        headers['mean_fivelen'] = {
-            'title': "Mean 5' primer length",
-            'description': "The mean 5' primer length in base pair",
-            'scale': 'RdYlGn',
+        headers["mean_fivelen"] = {
+            "title": "Mean 5' primer length",
+            "description": "The mean 5' primer length in base pair",
+            "scale": "RdYlGn",
         }
-        headers['std_fivelen'] = {
-            'title': "Std of 5' primer length",
-            'description': "The standard deviation of 5' primer length in base pair",
-            'scale': 'spectral',
+        headers["std_fivelen"] = {
+            "title": "Std of 5' primer length",
+            "description": "The standard deviation of 5' primer length in base pair",
+            "scale": "spectral",
         }
-        headers['max_fivelen'] = {
-            'title': "Max 5' primer length",
-            'description': "The maximum 5' primer length in base pair",
-            'scale': 'RdYlGn',
+        headers["max_fivelen"] = {
+            "title": "Max 5' primer length",
+            "description": "The maximum 5' primer length in base pair",
+            "scale": "RdYlGn",
         }
 
         config = {
@@ -189,38 +173,38 @@ class MultiqcModule(BaseMultiqcModule):
                     poly(A) length, and couple of primers detected for each CCS.
                     The table presents min, max, mean, standard deviation for each parameters.
                     """,
-            plot=table.plot(plot_data, headers, config)
+            plot=table.plot(plot_data, headers, config),
         )
 
     def add_section_threelen(self):
         plot_data = dict()
         for s_name, data in self.isoseq3_refine_data_csv.items():
             plot_data[s_name] = dict()
-            plot_data[s_name]['min_threelen'] = data['min_threelen']
-            plot_data[s_name]['mean_threelen'] = data['mean_threelen']
-            plot_data[s_name]['std_threelen'] = data['std_threelen']
-            plot_data[s_name]['max_threelen'] = data['max_threelen']
+            plot_data[s_name]["min_threelen"] = data["min_threelen"]
+            plot_data[s_name]["mean_threelen"] = data["mean_threelen"]
+            plot_data[s_name]["std_threelen"] = data["std_threelen"]
+            plot_data[s_name]["max_threelen"] = data["max_threelen"]
 
         headers = OrderedDict()
-        headers['min_threelen'] = {
-            'title': "Min 3' primer length",
-            'description': "The minimum 3' primer length in base pair",
-            'scale': 'spectral',
+        headers["min_threelen"] = {
+            "title": "Min 3' primer length",
+            "description": "The minimum 3' primer length in base pair",
+            "scale": "spectral",
         }
-        headers['mean_threelen'] = {
-            'title': "Mean 3' primer length",
-            'description': "The mean 3' primer length in base pair",
-            'scale': 'RdYlGn',
+        headers["mean_threelen"] = {
+            "title": "Mean 3' primer length",
+            "description": "The mean 3' primer length in base pair",
+            "scale": "RdYlGn",
         }
-        headers['std_threelen'] = {
-            'title': "Std of 3' primer length",
-            'description': "The standard deviation of 3' primer length in base pair",
-            'scale': 'spectral',
+        headers["std_threelen"] = {
+            "title": "Std of 3' primer length",
+            "description": "The standard deviation of 3' primer length in base pair",
+            "scale": "spectral",
         }
-        headers['max_threelen'] = {
-            'title': "Max 3' primer length",
-            'description': "The maximum 3' primer length in base pair",
-            'scale': 'RdYlGn',
+        headers["max_threelen"] = {
+            "title": "Max 3' primer length",
+            "description": "The maximum 3' primer length in base pair",
+            "scale": "RdYlGn",
         }
 
         config = {
@@ -237,38 +221,38 @@ class MultiqcModule(BaseMultiqcModule):
                     poly(A) length, and couple of primers detected for each CCS.
                     The table presents min, max, mean, standard deviation for each parameters.
                     """,
-            plot=table.plot(plot_data, headers, config)
+            plot=table.plot(plot_data, headers, config),
         )
 
     def add_section_polyalen(self):
         plot_data = dict()
         for s_name, data in self.isoseq3_refine_data_csv.items():
             plot_data[s_name] = dict()
-            plot_data[s_name]['min_polyAlen'] = data['min_polyAlen']
-            plot_data[s_name]['mean_polyAlen'] = data['mean_polyAlen']
-            plot_data[s_name]['std_polyAlen'] = data['std_polyAlen']
-            plot_data[s_name]['max_polyAlen'] = data['max_polyAlen']
+            plot_data[s_name]["min_polyAlen"] = data["min_polyAlen"]
+            plot_data[s_name]["mean_polyAlen"] = data["mean_polyAlen"]
+            plot_data[s_name]["std_polyAlen"] = data["std_polyAlen"]
+            plot_data[s_name]["max_polyAlen"] = data["max_polyAlen"]
 
             headers = OrderedDict()
-            headers['min_polyAlen'] = {
-                'title': "Min polyA tail length",
-                'description': "The minimum polyA tail length in base pair",
-                'scale': 'spectral',
+            headers["min_polyAlen"] = {
+                "title": "Min polyA tail length",
+                "description": "The minimum polyA tail length in base pair",
+                "scale": "spectral",
             }
-            headers['mean_polyAlen'] = {
-                'title': "Mean polyA tail length",
-                'description': "The mean polyA tail length in base pair",
-                'scale': 'RdYlGn',
+            headers["mean_polyAlen"] = {
+                "title": "Mean polyA tail length",
+                "description": "The mean polyA tail length in base pair",
+                "scale": "RdYlGn",
             }
-            headers['std_polyAlen'] = {
-                'title': "Std of polyA tail length",
-                'description': "The standard deviation of polyA tail length in base pair",
-                'scale': 'spectral',
+            headers["std_polyAlen"] = {
+                "title": "Std of polyA tail length",
+                "description": "The standard deviation of polyA tail length in base pair",
+                "scale": "spectral",
             }
-            headers['max_polyAlen'] = {
-                'title': "Max polyA tail length",
-                'description': "The maximum polyA tail length in base pair",
-                'scale': 'RdYlGn',
+            headers["max_polyAlen"] = {
+                "title": "Max polyA tail length",
+                "description": "The maximum polyA tail length in base pair",
+                "scale": "RdYlGn",
             }
 
             config = {
@@ -285,38 +269,38 @@ class MultiqcModule(BaseMultiqcModule):
                     poly(A) length, and couple of primers detected for each CCS.
                     The table presents min, max, mean, standard deviation for each parameters.
                     """,
-            plot=table.plot(plot_data, headers, config)
+            plot=table.plot(plot_data, headers, config),
         )
 
     def add_section_insertlen(self):
         plot_data = dict()
         for s_name, data in self.isoseq3_refine_data_csv.items():
             plot_data[s_name] = dict()
-            plot_data[s_name]['min_insertlen'] = data['min_insertlen']
-            plot_data[s_name]['mean_insertlen'] = data['mean_insertlen']
-            plot_data[s_name]['std_insertlen'] = data['std_insertlen']
-            plot_data[s_name]['max_insertlen'] = data['max_insertlen']
+            plot_data[s_name]["min_insertlen"] = data["min_insertlen"]
+            plot_data[s_name]["mean_insertlen"] = data["mean_insertlen"]
+            plot_data[s_name]["std_insertlen"] = data["std_insertlen"]
+            plot_data[s_name]["max_insertlen"] = data["max_insertlen"]
 
             headers = OrderedDict()
-            headers['min_insertlen'] = {
-                'title': "Min insert length",
-                'description': "The minimum insert length in base pair",
-                'scale': 'spectral',
+            headers["min_insertlen"] = {
+                "title": "Min insert length",
+                "description": "The minimum insert length in base pair",
+                "scale": "spectral",
             }
-            headers['mean_insertlen'] = {
-                'title': "Mean insert length",
-                'description': "The mean insert length in base pair",
-                'scale': 'RdYlGn',
+            headers["mean_insertlen"] = {
+                "title": "Mean insert length",
+                "description": "The mean insert length in base pair",
+                "scale": "RdYlGn",
             }
-            headers['std_insertlen'] = {
-                'title': "Std of insert length",
-                'description': "The standard deviation of insert length in base pair",
-                'scale': 'spectral',
+            headers["std_insertlen"] = {
+                "title": "Std of insert length",
+                "description": "The standard deviation of insert length in base pair",
+                "scale": "spectral",
             }
-            headers['max_insertlen'] = {
-                'title': "Max insert length",
-                'description': "The maximum insert length in base pair",
-                'scale': 'RdYlGn',
+            headers["max_insertlen"] = {
+                "title": "Max insert length",
+                "description": "The maximum insert length in base pair",
+                "scale": "RdYlGn",
             }
 
             config = {
@@ -333,5 +317,5 @@ class MultiqcModule(BaseMultiqcModule):
                     poly(A) length, and couple of primers detected for each CCS.
                     The table presents min, max, mean, standard deviation for each parameters.
                     """,
-            plot=table.plot(plot_data, headers, config)
+            plot=table.plot(plot_data, headers, config),
         )